Extract CallbackPlot class from BokehPlot
so that we can reuse it for the plotly backend
